Fix import cycle related to EnrichedStopPlace changes
