+++ ui/src/components/map/stop-areas/StopAreaPopup.tsx
@@ -1,10 +1,10 @@
 import { useTranslation } from 'react-i18next';
 import { MdDelete } from 'react-icons/md';
 import { Popup } from 'react-map-gl/maplibre';
-import { EnrichedStopPlace } from '../../../hooks';
 import { Column, Row } from '../../../layoutComponents';
 import { Path, routeDetails } from '../../../router/routeDetails';
 import { parseDate } from '../../../time';
+import { EnrichedStopPlace } from '../../../types';
 import { CloseIconButton, SimpleButton } from '../../../uiComponents';
 import {
   findKeyValueParsed,
