+++ ui/src/components/stop-registry/stop-areas/stop-area-details/components/StopAreaDetailsEdit.tsx
@@ -3,10 +3,11 @@ import React, { ForwardRefRenderFunction, forwardRef, useMemo } from 'react';
 import { FormProvider, UseFormReturn, useForm } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import { twMerge } from 'tailwind-merge';
-import { EnrichedStopPlace, useLoader } from '../../../../../hooks';
+import { useLoader } from '../../../../../hooks';
 import { Column } from '../../../../../layoutComponents';
 import { Operation } from '../../../../../redux';
 import { mapToISODate } from '../../../../../time';
+import { EnrichedStopPlace } from '../../../../../types';
 import { mapLngLatToPoint, showSuccessToast } from '../../../../../utils';
 import {
   FormColumn,
