+++ ui/src/utils/stop-registry/stopPlace.ts
@@ -13,6 +13,10 @@ import {
   StopRegistryStopPlaceInput,
 } from '../../generated/graphql';
 import { hasTypeName } from '../../graphql';
+import {
+  QuayEnrichmentProperties,
+  StopPlaceEnrichmentProperties,
+} from '../../types';
 import { StopPlaceState } from '../../types/stop-registry';
 import { findKeyValue, findKeyValueParsed } from '../findKeyValue';
 import { mapLngLatToPoint } from '../gis';
@@ -51,40 +55,6 @@ export const getStopPlacesFromQueryResult = <T extends StopPlaceType>(
 
 // Required in DB so can't be null.
 export const defaultAccessibilityLevel = StopRegistryAccessibilityLevel.Unknown;
-
-export type StopPlaceEnrichmentProperties = {
-  readonly nameSwe?: string | undefined;
-  readonly nameLongFin?: string | undefined;
-  readonly nameLongSwe?: string | undefined;
-  readonly abbreviationFin?: string | undefined;
-  readonly abbreviationSwe?: string | undefined;
-  readonly abbreviation5CharFin?: string | undefined;
-  readonly abbreviation5CharSwe?: string | undefined;
-  readonly name?: string | undefined;
-  readonly municipality?: string | undefined;
-  readonly fareZone?: string | undefined;
-  readonly locationLat?: number | undefined;
-  readonly locationLong?: number | undefined;
-  readonly validityStart?: string | undefined;
-  readonly validityEnd?: string | undefined;
-};
-
-export type QuayEnrichmentProperties = {
-  readonly elyNumber: string | null;
-  readonly locationFin: string | null;
-  readonly locationSwe: string | null;
-  readonly streetAddress: string | null;
-  readonly postalCode: string | null;
-  readonly functionalArea: number | null;
-  readonly stopState: StopPlaceState | null;
-  readonly accessibilityLevel: StopRegistryAccessibilityLevel;
-  readonly stopType: {
-    readonly mainLine: boolean;
-    readonly virtual: boolean;
-    readonly interchange: boolean;
-    readonly railReplacement: boolean;
-  };
-};
 
 const findAlternativeName = (
   stopPlace: Pick<StopRegistryStopPlace, 'alternativeNames'>,
